Added IAuthentication to authenticate user and refactored create/delete post

+++ officey/src/main/java/com/example/officey/service/PostServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.officey.service;
 
+import com.example.officey.controller.SecurityController;
 import com.example.officey.model.Post;
 import com.example.officey.model.User;
 import com.example.officey.repository.PostRepository;
@@ -7,6 +8,8 @@ import com.example.officey.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+
+import java.net.Authenticator;
 
 @Service
 public class PostServiceImpl implements PostService {
@@ -16,8 +19,12 @@ public class PostServiceImpl implements PostService {
     @Autowired
     UserRepository userRepository;
 
+    @Autowired
+    SecurityController securityController;
+
     @Override
-    public Post createPostInDB(Post post, String username) {
+    public Post createPostInDB(Post post) {
+        String username = securityController.getCurrentUserName();
         User user = userRepository.findByUsername(username);
         post.setUser(user);
         return postRepository.save(post);
@@ -28,6 +35,12 @@ public class PostServiceImpl implements PostService {
         return postRepository.findAll();
     }
 
+    public Iterable<Post> listPostsOfUser() {
+        String username = securityController.getCurrentUserName();
+        Long userId = userRepository.findByUsername(username).getId();
+        return postRepository.findPostsByUserId(userId);
+    }
+
 //    @Override
 //    public Post getSinglePost(Long postId) {
 //        return postRepository.findPostById(postId);
@@ -35,7 +48,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public String deletePostByIdInDB(Long postId) {
-        postRepository.deleteById(postId);
-        return "Post has been deleted";
+
+        String username = securityController.getCurrentUserName();
+
+        if(postRepository.findById(postId).get().getUser().getUsername().equals(username)) {
+            postRepository.deleteById(postId);
+            return "Post has been deleted";
+        } else {
+            return "THIS IS NOT YOUR POST TO DELETE";
+        }
+
     }
 }
